Remove DC functions from Mesh class
