+++ libfive/src/libfive.cpp
@@ -16,7 +16,7 @@ You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "libfive/render/brep/region.hpp"
 #include "libfive/render/brep/contours.hpp"
-#include "libfive/render/brep/dc/mesh.hpp"
+#include "libfive/render/brep/mesh.hpp"
 
 #include "libfive/render/discrete/voxels.hpp"
 #include "libfive/render/discrete/heightmap.hpp"
