+++ libfive/src/render/brep/dc/dc_mesher.cpp
@@ -1,7 +1,7 @@
 /*
 libfive: a CAD kernel for modeling with implicit functions
 
-Copyright (C) 2017  Matt Keeter
+Copyright (C) 2018  Matt Keeter
 
 This Source Code Form is subject to the terms of the Mozilla Public
 License, v. 2.0. If a copy of the MPL was not distributed with this file,
@@ -11,8 +11,8 @@ You can obtain one at http://mozilla.org/MPL/2.0/.
 #include <fstream>
 #include <boost/algorithm/string/predicate.hpp>
 
+#include "libfive/render/brep/dc/dc_mesher.hpp"
 #include "libfive/render/brep/dual.hpp"
-#include "libfive/render/brep/dc/mesh.hpp"
 #include "libfive/render/brep/dc/xtree_pool.hpp"
 #include "libfive/render/brep/dc/intersection_aligner.hpp"
 
@@ -22,10 +22,8 @@ You can obtain one at http://mozilla.org/MPL/2.0/.
 
 namespace Kernel {
 
-const float Mesh::MAX_PROGRESS = 3.0f;
-
 template <Axis::Axis A>
-void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
+void DCMesher::load(const std::array<const XTree<3>*, 4>& ts)
 {
     // Sanity-checking that all cells have a Leaf struct allocated
     for (auto& t : ts)
@@ -77,7 +75,7 @@ void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
 }
 
 template <Axis::Axis A, bool D>
-void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
+void DCMesher::load(const std::array<const XTree<3>*, 4>& ts)
 {
     int es[4];
     {   // Unpack edge vertex pairs into edge indices
@@ -113,9 +111,9 @@ void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
 
         if (ts[i]->leaf->index[vi] == 0)
         {
-            ts[i]->leaf->index[vi] = verts.size();
+            ts[i]->leaf->index[vi] = m.verts.size();
 
-            verts.push_back(ts[i]->vert(vi).template cast<float>());
+            m.verts.push_back(ts[i]->vert(vi).template cast<float>());
         }
         vs[i] = ts[i]->leaf->index[vi];
     }
@@ -171,10 +169,6 @@ void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
     // Pick a triangulation that prevents triangles from folding back
     // on each other by checking normals.
     std::array<Eigen::Vector3f, 4> norms;
-    auto saveNorm = [&](int a, int b, int c){
-        norms[a] = (verts[vs[b]] - verts[vs[a]]).cross
-                   (verts[vs[c]] - verts[vs[a]]).normalized();
-    };
 
     // Computes and saves a corner normal.  a,b,c must be right-handed
     // according to the quad winding, which looks like
@@ -182,6 +176,10 @@ void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
     //     |         |
     //     |         |
     //     0---------1
+    auto saveNorm = [&](int a, int b, int c){
+        norms[a] = (m.verts[vs[b]] - m.verts[vs[a]]).cross
+                   (m.verts[vs[c]] - m.verts[vs[a]]).normalized();
+    };
     saveNorm(0, 1, 2);
     saveNorm(1, 3, 0);
     saveNorm(2, 0, 3);
@@ -191,7 +189,7 @@ void Mesh::load(const std::array<const XTree<3>*, 4>& ts)
     auto push_triangle = [&](uint32_t a, uint32_t b, uint32_t c) {
         if (a != b && b != c && a != c)
         {
-            branes.push_back({a, b, c});
+            m.branes.push_back({a, b, c});
         }
     };
 
@@ -451,51 +449,7 @@ void Mesh::checkAndAddTriangle(const XTree<3>* a, const XTree<3>* b,
     branes.push_back({ aIndex, bIndex, intersectionIndex });
 }
 
-////////////////////////////////////////////////////////////////////////////////
-
-std::unique_ptr<Mesh> Mesh::render(const Tree t, const Region<3>& r,
-                                   double min_feature, double max_err,
-                                   bool multithread,
-                                   ProgressCallback progress_callback)
-{
-    std::atomic_bool cancel(false);
-    std::map<Tree::Id, float> vars;
-    return render(t, vars, r, min_feature, max_err,
-                  multithread ? 8 : 1, cancel, progress_callback);
-}
-
-std::unique_ptr<Mesh> Mesh::render(
-            const Tree t, const std::map<Tree::Id, float>& vars,
-            const Region<3>& r, double min_feature, double max_err,
-            unsigned workers, std::atomic_bool& cancel,
-            ProgressCallback progress_callback)
-{
-    std::vector<XTreeEvaluator, Eigen::aligned_allocator<XTreeEvaluator>> es;
-    es.reserve(workers);
-    for (unsigned i=0; i < workers; ++i)
-    {
-        es.emplace_back(XTreeEvaluator(t, vars));
-    }
-
-    return render(es.data(), r, min_feature, max_err, workers, cancel,
-                  progress_callback);
-}
-
-std::unique_ptr<Mesh> Mesh::render(
-        XTreeEvaluator* es,
-        const Region<3>& r, double min_feature, double max_err,
-        int workers, std::atomic_bool& cancel,
-        ProgressCallback progress_callback)
-{
-    auto t = XTreePool<3>::build(es, r, min_feature, max_err, workers, cancel,
-                                 progress_callback);
-    auto out = mesh(t, cancel, progress_callback);
-    t.reset(progress_callback);
-
-    return out;
-}
-
-std::unique_ptr<Mesh> Mesh::mesh(const Root<XTree<3>, XTree<3>::Leaf>& xtree,
+std::unique_ptr<Mesh> DCMesher::mesh(const Root<XTree<3>, XTree<3>::Leaf>& xtree,
                                  std::atomic_bool& cancel,
                                  ProgressCallback progress_callback)
 {
@@ -512,6 +466,7 @@ std::unique_ptr<Mesh> Mesh::mesh(const Root<XTree<3>, XTree<3>::Leaf>& xtree,
 
     // Perform marching squares
     auto m = std::unique_ptr<Mesh>(new Mesh());
+    DCMesher d(*m);
 
     if (cancel.load() || xtree.get() == nullptr)
     {
@@ -524,7 +479,7 @@ std::unique_ptr<Mesh> Mesh::mesh(const Root<XTree<3>, XTree<3>::Leaf>& xtree,
                 xtree.size(), 1.0f,
                 progress_callback, done, cancel);
 
-        Dual<3>::walk(xtree.get(), *m, progress_watcher);
+        Dual<3>::walk(xtree.get(), d, progress_watcher);
 
         done.store(true);
         delete progress_watcher;
@@ -555,80 +510,6 @@ std::unique_ptr<Mesh> Mesh::mesh(const Root<XTree<3>, XTree<3>::Leaf>& xtree,
 #endif
         return m;
     }
-}
-
-void Mesh::line(const Eigen::Vector3f& a, const Eigen::Vector3f& b)
-{
-    uint32_t a_ = verts.size();
-    verts.push_back(a);
-    uint32_t b_ = verts.size();
-    verts.push_back(b);
-
-    branes.push_back({a_, a_, b_});
-}
-
-////////////////////////////////////////////////////////////////////////////////
-
-bool Mesh::saveSTL(const std::string& filename,
-                   const std::list<const Mesh*>& meshes)
-{
-    if (!boost::algorithm::iends_with(filename, ".stl"))
-    {
-        std::cerr << "Mesh::saveSTL: filename \"" << filename
-                  << "\" does not end in .stl" << std::endl;
-    }
-    std::ofstream file;
-    file.open(filename, std::ios::out | std::ios::binary);
-    if (!file.is_open())
-    {
-        std::cout << "Mesh::saveSTL: could not open " << filename
-                  << std::endl;
-        return false;
-    }
-
-    // File header (giving human-readable info about file type)
-    std::string header = "This is a binary STL exported from libfive.";
-    file.write(header.c_str(), header.length());
-
-    // Pad the rest of the header to 80 bytes
-    for (int i=header.length(); i < 80; ++i)
-    {
-        file.put(' ');
-    }
-
-    // Write the triangle count to the file
-    uint32_t num = std::accumulate(meshes.begin(), meshes.end(), (uint32_t)0,
-            [](uint32_t i, const Mesh* m){ return i + m->branes.size(); });
-    file.write(reinterpret_cast<char*>(&num), sizeof(num));
-
-    for (const auto& m : meshes)
-    {
-        for (const auto& t : m->branes)
-        {
-            // Write out the normal vector for this face (all zeros)
-            float norm[3] = {0, 0, 0};
-            file.write(reinterpret_cast<char*>(&norm), sizeof(norm));
-
-            // Iterate over vertices (which are indices into the verts list)
-            for (unsigned i=0; i < 3; ++i)
-            {
-                auto v = m->verts[t[i]];
-                float vert[3] = {v.x(), v.y(), v.z()};
-                file.write(reinterpret_cast<char*>(&vert), sizeof(vert));
-            }
-
-            // Write out this face's attribute short
-            uint16_t attrib = 0;
-            file.write(reinterpret_cast<char*>(&attrib), sizeof(attrib));
-        }
-    }
-
-    return true;
-}
-
-bool Mesh::saveSTL(const std::string& filename)
-{
-    return saveSTL(filename, {this});
 }
 
 }   // namespace Kernel
