+++ libfive/test/oracle.cpp
@@ -10,7 +10,7 @@ You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #include "catch.hpp"
 
-#include "libfive/render/brep/dc/mesh.hpp"
+#include "libfive/render/brep/mesh.hpp"
 #include "libfive/render/brep/region.hpp"
 
 #include "libfive/oracle/oracle_storage.hpp"
