+++ studio/include/studio/shape.hpp
@@ -25,7 +25,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include <QOpenGLFunctions>
 
 #include "libfive/tree/tree.hpp"
-#include "libfive/render/brep/dc/mesh.hpp"
+#include "libfive/render/brep/mesh.hpp"
 #include "libfive/eval/eval_jacobian.hpp"
 
 namespace Kernel { class Tape; /*  forward declaration */ }
